fixes for platforms where qreal is not a double btw, wouldn't it be cleaner if all those qMin(0.0, qMax(1.0, foo)) calls are calls to qBound instead?
svn path=/trunk/KDE/kdebase/workspace/libs/plasma/; revision=748283

+++ layouts/layoutanimator.cpp
@@ -220,9 +220,9 @@ void LayoutAnimator::moveEffectUpdateItem( qreal value , LayoutItem* item , Effe
     Widget* widget = dynamic_cast<Widget*>(item->graphicsItem());
 
     if ( widget && effect == FadeInMoveEffect )
-        widget->setOpacity( qMin(1.0,widget->opacity()+d->delta(value)) );
+        widget->setOpacity( qMin(qreal(1.0),widget->opacity()+d->delta(value)) );
     else if ( widget && effect == FadeOutMoveEffect )
-        widget->setOpacity( qMax(0.0,widget->opacity()-d->delta(value)) );
+        widget->setOpacity( qMax(qreal(0.0),widget->opacity()-d->delta(value)) );
 
     
     if ( effect == FadeInMoveEffect ) {
